+++ test/fixtures/projects/a-ts-project/tsconfig.json
@@ -34,6 +34,7 @@
       "resolveJsonModule": true
   },
   "include": [
-      "src"
+      "src",
+      "test"
   ]
 }
